add edit and remove functionality
