+++ src/style.css
@@ -1,3 +1,34 @@
 .item-details {
   display: flex;
+  flex-direction: row;
+}
+
+.clear > h3 {
+  text-align: center;
+}
+
+body {
+  background-color: #f5f5f5;
+}
+
+.whole-body {
+  border: 1px solid #ececec;
+  width: 50%;
+  margin: auto;
+  padding: 20px;
+}
+
+.books-container {
+  width: 50%;
+  margin: auto;
+}
+
+form {
+  text-align: center;
+}
+
+.button-details {
+  height: 20px;
+  margin: auto;
+  float: right;
 }
